move users domain tests to new package
