+++ src/test/java/pl/zdna/gcconnect/users/EmailVTest.java
@@ -1,4 +1,4 @@
-package pl.zdna.gcconnect.vgn;
+package pl.zdna.gcconnect.users;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -6,6 +6,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import pl.zdna.gcconnect.vgn.EmailValidator;
+import pl.zdna.gcconnect.vgn.Validator;
 
 public class EmailVTest {
     private final Validator<String> emailValidator = new EmailValidator();
